Scheduled ingestion, and the macOS trap that silently breaks it
Proved the weekly pipeline on 2025, since 2026 cannot prove it -- week one has
not been played. All four sources pull and reprice the real board off real
results:

  player_stats  19,421   depth_charts 554,215   injuries 6,068   snaps 26,612

  Nacua      27.4 ppg x4   ros 191 -> 251  (w=55%)
  A.J. Brown  8.8 ppg x4   ros 161 -> 123  (w=55%)
  Cook       22.8 ppg x4   ros 156 -> 211  (w=66%)

Proportionate, not hysterical, and the weights are the per-position ones.

Three layers so the data is current whenever you look:
  scripts/sync.py            what a scheduler runs; exit 0 fresh, 1 got
                             nothing, 2 crashed -- a scheduler is the only
                             thing watching, so the codes have to mean something
  in-app keeper thread       every 6h while the app is open, no-ops inside the
                             weekly window. A laptop is asleep most of the week,
                             so this is the layer that actually fires
  install_sync_job.sh        launchd, Tuesday 06:00 -- after MNF settles.
                             launchd not cron: cron loses the run if the
                             machine is asleep at that instant, launchd runs it
                             on wake

THE TRAP. Installing that job appeared to work -- launchctl accepted it,
listed it, reported runs=1. It had never executed. macOS refuses to let a
background agent run out of Desktop, Documents or Downloads, and this repo is
in Downloads. It does not prompt and it does not log: exit 78 EX_CONFIG, empty
log file, nothing in stderr. Isolated it with a control -- identical plist
pointed at /bin/date exits 0, pointed at the venv here exits 78.

So the installer detects a protected path and refuses, with both fixes, rather
than installing something that quietly never runs. The broken job is removed.

data/logs/ was not gitignored; scheduler output can echo league ids.

+++ fantasyedge/models/inseason.py
@@ -182,7 +182,10 @@ def observe(weekly: pl.DataFrame, season: int, through_week: int) -> pl.DataFram
         return pl.DataFrame()
 
     agg = [
-        pl.len().alias("games"),
+        # Float64, not the UInt32 `len` hands back. Nothing here subtracts two
+        # counts, but this project has shipped five separate underflow bugs
+        # from exactly this dtype and the cast costs nothing.
+        pl.len().cast(pl.Float64).alias("games"),
         pl.col("fantasy_points_ppr").mean().alias("ppg"),
     ]
     if "opportunity" in d.columns and "points_per_opp" in d.columns:
